+++ test/small1/testharness.h
@@ -12,6 +12,6 @@
 extern void exit(int);
 
 /* Always call E with a non-zero number */
-#define E(n) { printf("Error %d\n", n); exit(n); }
+#define E(n) { printf("Error %d\n", (int)n); exit(n); }
 #define SUCCESS { printf("Success\n"); exit(0); }
